pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-novel-writer"
version = "0.1.0"
description = "An AI-powered novel writing assistant using OpenAI"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
keywords = ["ai", "novel", "writing", "openai", "gpt"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
novel-writer = "novel_writer.main:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["novel_writer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
